+++ src/main/java/com/contentful/java/cma/ModuleContentTypes.java
@@ -28,6 +28,8 @@ import java.util.concurrent.Executor;
 
 import retrofit2.Retrofit;
 
+import static com.contentful.java.cma.Constants.DEFAULT_ENVIRONMENT;
+
 /**
  * Content Types Module.
  */
@@ -56,10 +58,35 @@ public final class ModuleContentTypes extends AbsModule<ServiceContentTypes> {
    * @throws IllegalArgumentException if contentType is null.
    */
   @SuppressWarnings("unchecked")
-  public CMAContentType create(String spaceId,
-                               CMAContentType contentType) {
+  public CMAContentType create(
+      String spaceId,
+      CMAContentType contentType) {
+    return create(spaceId, DEFAULT_ENVIRONMENT, contentType);
+  }
+
+  /**
+   * Create a new Content Type in an Environment.
+   * In case the given {@code contentType} has an ID associated with it, that ID will be used,
+   * otherwise the server will auto-generate an ID that will be contained in the response upon
+   * success.
+   *
+   * @param spaceId       Space ID
+   * @param environmentId Environment ID
+   * @param contentType   Content Type
+   * @return {@link CMAContentType} result instance
+   * @throws IllegalArgumentException if spaceId is null.
+   * @throws IllegalArgumentException if environmentId is null.
+   * @throws IllegalArgumentException if contentType is null.
+   */
+  @SuppressWarnings("unchecked")
+  public CMAContentType create(
+      String spaceId,
+      String environmentId,
+      CMAContentType contentType) {
+
     assertNotNull(spaceId, "spaceId");
     assertNotNull(contentType, "contentType");
+    assertNotNull(environmentId, "environmentId");
 
     final String contentTypeId = contentType.getId();
 
@@ -68,9 +95,9 @@ public final class ModuleContentTypes extends AbsModule<ServiceContentTypes> {
 
     try {
       if (contentTypeId == null) {
-        return service.create(spaceId, contentType).blockingFirst();
+        return service.create(spaceId, environmentId, contentType).blockingFirst();
       } else {
-        return service.create(spaceId, contentTypeId, contentType).blockingFirst();
+        return service.create(spaceId, environmentId, contentTypeId, contentType).blockingFirst();
       }
     } finally {
       contentType.setSystem(sys);
@@ -87,9 +114,25 @@ public final class ModuleContentTypes extends AbsModule<ServiceContentTypes> {
    * @throws IllegalArgumentException if contentTypeId is null.
    */
   public Integer delete(String spaceId, String contentTypeId) {
+    return delete(spaceId, DEFAULT_ENVIRONMENT, contentTypeId);
+  }
+
+  /**
+   * Delete a Content Type.
+   *
+   * @param spaceId       Space ID
+   * @param environmentId Environment ID
+   * @param contentTypeId Content Type ID
+   * @return Integer representing the result of the operation
+   * @throws IllegalArgumentException if spaceId is null.
+   * @throws IllegalArgumentException if environmentId is null.
+   * @throws IllegalArgumentException if contentTypeId is null.
+   */
+  public Integer delete(String spaceId, String environmentId, String contentTypeId) {
     assertNotNull(spaceId, "spaceId");
+    assertNotNull(environmentId, "environmentId");
     assertNotNull(contentTypeId, "contentTypeId");
-    return service.delete(spaceId, contentTypeId).blockingFirst().code();
+    return service.delete(spaceId, environmentId, contentTypeId).blockingFirst().code();
   }
 
   /**
@@ -101,9 +144,11 @@ public final class ModuleContentTypes extends AbsModule<ServiceContentTypes> {
    * @throws IllegalArgumentException if contentTypeId is null.
    */
   public Integer delete(CMAContentType contentType) {
-    assertNotNull(contentType.getSpaceId(), "spaceId");
-    assertNotNull(contentType.getId(), "contentTypeId");
-    return service.delete(contentType.getSpaceId(), contentType.getId()).blockingFirst().code();
+    final String spaceId = getSpaceIdOrThrow(contentType, "contentType");
+    final String environmentId = contentType.getEnvironmentId();
+    final String contentTypeId = getResourceIdOrThrow(contentType, "contentType");
+
+    return delete(spaceId, environmentId, contentTypeId);
   }
 
   /**
@@ -116,7 +161,21 @@ public final class ModuleContentTypes extends AbsModule<ServiceContentTypes> {
    * @throws IllegalArgumentException if spaceId is null.
    */
   public CMAArray<CMAContentType> fetchAll(String spaceId) {
-    return fetchAll(spaceId, new HashMap<String, String>());
+    return fetchAll(spaceId, new HashMap<>());
+  }
+
+  /**
+   * Fetch all Content Types from an Environment, using default query parameter.
+   * <p>
+   * This fetch uses the default parameter defined in {@link DefaultQueryParameter#FETCH}
+   *
+   * @param spaceId       Space ID
+   * @param environmentId Environment ID
+   * @return {@link CMAArray} result instance
+   * @throws IllegalArgumentException if spaceId is null.
+   */
+  public CMAArray<CMAContentType> fetchAll(String spaceId, String environmentId) {
+    return fetchAll(spaceId, environmentId, new HashMap<>());
   }
 
   /**
@@ -128,9 +187,26 @@ public final class ModuleContentTypes extends AbsModule<ServiceContentTypes> {
    * @throws IllegalArgumentException if spaceId is null.
    */
   public CMAArray<CMAContentType> fetchAll(String spaceId, Map<String, String> query) {
+    return fetchAll(spaceId, DEFAULT_ENVIRONMENT, query);
+  }
+
+  /**
+   * Fetch all Content Types from a Space with query parameters.
+   *
+   * @param spaceId       Space ID
+   * @param environmentId Environment ID
+   * @param query         Query to narrow down the content_types to be searched for
+   * @return {@link CMAArray} result instance
+   * @throws IllegalArgumentException if spaceId is null.
+   * @throws IllegalArgumentException if environmentId is null.
+   */
+  public CMAArray<CMAContentType> fetchAll(
+      String spaceId,
+      String environmentId,
+      Map<String, String> query) {
     assertNotNull(spaceId, "spaceId");
     DefaultQueryParameter.putIfNotSet(query, DefaultQueryParameter.FETCH);
-    return service.fetchAll(spaceId, query).blockingFirst();
+    return service.fetchAll(spaceId, environmentId, query).blockingFirst();
   }
 
   /**
@@ -145,7 +221,25 @@ public final class ModuleContentTypes extends AbsModule<ServiceContentTypes> {
   public CMAContentType fetchOne(String spaceId, String contentTypeId) {
     assertNotNull(spaceId, "spaceId");
     assertNotNull(contentTypeId, "contentTypeId");
-    return service.fetchOne(spaceId, contentTypeId).blockingFirst();
+    return fetchOne(spaceId, DEFAULT_ENVIRONMENT, contentTypeId);
+  }
+
+  /**
+   * Fetch a Content Type with the given {@code contentTypeId} from an Environment.
+   *
+   * @param spaceId       Space ID
+   * @param environmentId Environment ID
+   * @param contentTypeId Content Type ID
+   * @return {@link CMAContentType} result instance
+   * @throws IllegalArgumentException if spaceId is null.
+   * @throws IllegalArgumentException if environmentId is null.
+   * @throws IllegalArgumentException if contentTypeId is null.
+   */
+  public CMAContentType fetchOne(String spaceId, String environmentId, String contentTypeId) {
+    assertNotNull(spaceId, "spaceId");
+    assertNotNull(environmentId, "environmentId");
+    assertNotNull(contentTypeId, "contentTypeId");
+    return service.fetchOne(spaceId, environmentId, contentTypeId).blockingFirst();
   }
 
   /**
@@ -162,10 +256,12 @@ public final class ModuleContentTypes extends AbsModule<ServiceContentTypes> {
 
     final String contentTypeId = getResourceIdOrThrow(contentType, "contentType");
     final String spaceId = getSpaceIdOrThrow(contentType, "contentType");
+    final String environmentId = contentType.getEnvironmentId();
 
     return service.publish(
         contentType.getVersion(),
         spaceId,
+        environmentId,
         contentTypeId
     ).blockingFirst();
   }
@@ -184,8 +280,9 @@ public final class ModuleContentTypes extends AbsModule<ServiceContentTypes> {
 
     final String contentTypeId = getResourceIdOrThrow(contentType, "contentType");
     final String spaceId = getSpaceIdOrThrow(contentType, "contentType");
+    final String environmentId = contentType.getEnvironmentId();
 
-    return service.unPublish(spaceId, contentTypeId).blockingFirst();
+    return service.unPublish(spaceId, environmentId, contentTypeId).blockingFirst();
   }
 
   /**
@@ -205,6 +302,7 @@ public final class ModuleContentTypes extends AbsModule<ServiceContentTypes> {
 
     final String contentTypeId = getResourceIdOrThrow(contentType, "contentType");
     final String spaceId = getSpaceIdOrThrow(contentType, "contentType");
+    final String environmentId = contentType.getEnvironmentId();
     final Integer version = getVersionOrThrow(contentType, "update");
 
     final CMASystem system = contentType.getSystem();
@@ -214,6 +312,7 @@ public final class ModuleContentTypes extends AbsModule<ServiceContentTypes> {
       return service.update(
           version,
           spaceId,
+          environmentId,
           contentTypeId,
           contentType
       ).blockingFirst();
@@ -236,8 +335,9 @@ public final class ModuleContentTypes extends AbsModule<ServiceContentTypes> {
 
     final String contentTypeId = getResourceIdOrThrow(contentType, "contentType");
     final String spaceId = getSpaceIdOrThrow(contentType, "contentType");
+    final String environmentId = contentType.getEnvironmentId();
 
-    return service.fetchAllSnapshots(spaceId, contentTypeId).blockingFirst();
+    return service.fetchAllSnapshots(spaceId, environmentId, contentTypeId).blockingFirst();
   }
 
   /**
@@ -258,8 +358,14 @@ public final class ModuleContentTypes extends AbsModule<ServiceContentTypes> {
 
     final String contentTypeId = getResourceIdOrThrow(contentType, "contentType");
     final String spaceId = getSpaceIdOrThrow(contentType, "contentType");
+    final String environmentId = contentType.getEnvironmentId();
 
-    return service.fetchOneSnapshot(spaceId, contentTypeId, snapshotId).blockingFirst();
+    return service.fetchOneSnapshot(
+        spaceId,
+        environmentId,
+        contentTypeId,
+        snapshotId
+    ).blockingFirst();
   }
 
   /**
@@ -287,12 +393,41 @@ public final class ModuleContentTypes extends AbsModule<ServiceContentTypes> {
      * @throws IllegalArgumentException if contentType is null.
      * @throws IllegalArgumentException if contentTypeId is null.
      */
-    public CMACallback<CMAContentType> create(final String spaceId,
-                                              final CMAContentType contentType,
-                                              CMACallback<CMAContentType> callback) {
+    public CMACallback<CMAContentType> create(
+        final String spaceId,
+        final CMAContentType contentType,
+        CMACallback<CMAContentType> callback) {
       return defer(new DefFunc<CMAContentType>() {
         @Override CMAContentType method() {
           return ModuleContentTypes.this.create(spaceId, contentType);
+        }
+      }, callback);
+    }
+
+    /**
+     * Create a new Content Type in an environment.
+     * In case the given {@code contentType} has an ID associated with it, that ID will be used,
+     * otherwise the server will auto-generate an ID that will be contained in the response upon
+     * success.
+     *
+     * @param spaceId       Space ID
+     * @param environmentId Environment ID
+     * @param contentType   Content Type
+     * @param callback      Callback
+     * @return the given {@code CMACallback} instance
+     * @throws IllegalArgumentException if spaceId is null.
+     * @throws IllegalArgumentException if environmentId is null.
+     * @throws IllegalArgumentException if contentType is null.
+     * @throws IllegalArgumentException if contentTypeId is null.
+     */
+    public CMACallback<CMAContentType> create(
+        final String spaceId,
+        final String environmentId,
+        final CMAContentType contentType,
+        CMACallback<CMAContentType> callback) {
+      return defer(new DefFunc<CMAContentType>() {
+        @Override CMAContentType method() {
+          return ModuleContentTypes.this.create(spaceId, environmentId, contentType);
         }
       }, callback);
     }
@@ -313,6 +448,29 @@ public final class ModuleContentTypes extends AbsModule<ServiceContentTypes> {
       return defer(new DefFunc<Integer>() {
         @Override Integer method() {
           return ModuleContentTypes.this.delete(spaceId, contentTypeId);
+        }
+      }, callback);
+    }
+
+    /**
+     * Delete a Content Type from an environment.
+     *
+     * @param spaceId       Space ID
+     * @param environmentId Environment ID
+     * @param contentTypeId Content Type ID
+     * @param callback      Callback
+     * @return the given {@code CMACallback} instance
+     * @throws IllegalArgumentException if spaceId is null.
+     * @throws IllegalArgumentException if contentTypeId is null.
+     */
+    public CMACallback<Integer> delete(
+        final String spaceId,
+        final String environmentId,
+        final String contentTypeId,
+        CMACallback<Integer> callback) {
+      return defer(new DefFunc<Integer>() {
+        @Override Integer method() {
+          return ModuleContentTypes.this.delete(spaceId, environmentId, contentTypeId);
         }
       }, callback);
     }
@@ -343,10 +501,34 @@ public final class ModuleContentTypes extends AbsModule<ServiceContentTypes> {
      * @return the given {@code CMACallback} instance
      * @throws IllegalArgumentException if spaceId is null.
      */
-    public CMACallback<CMAArray<CMAContentType>> fetchAll(final String spaceId,
-                                                          CMACallback<
-                                                              CMAArray<CMAContentType>> callback) {
-      return fetchAll(spaceId, new HashMap<String, String>(), callback);
+    public CMACallback<CMAArray<CMAContentType>> fetchAll(
+        final String spaceId,
+        CMACallback<CMAArray<CMAContentType>> callback) {
+      return defer(new DefFunc<CMAArray<CMAContentType>>() {
+        @Override CMAArray<CMAContentType> method() {
+          return ModuleContentTypes.this.fetchAll(spaceId);
+        }
+      }, callback);
+    }
+
+    /**
+     * Fetch all Content Types from an Environment.
+     *
+     * @param spaceId       Space ID
+     * @param environmentId Environment ID
+     * @param callback      Callback
+     * @return the given {@code CMACallback} instance
+     * @throws IllegalArgumentException if spaceId is null.
+     */
+    public CMACallback<CMAArray<CMAContentType>> fetchAll(
+        final String spaceId,
+        final String environmentId,
+        CMACallback<CMAArray<CMAContentType>> callback) {
+      return defer(new DefFunc<CMAArray<CMAContentType>>() {
+        @Override CMAArray<CMAContentType> method() {
+          return ModuleContentTypes.this.fetchAll(spaceId, environmentId);
+        }
+      }, callback);
     }
 
     /**
@@ -358,13 +540,12 @@ public final class ModuleContentTypes extends AbsModule<ServiceContentTypes> {
      * @return the given {@code CMACallback} instance
      * @throws IllegalArgumentException if spaceId is null.
      */
-    public CMACallback<CMAArray<CMAContentType>> fetchAll(final String spaceId,
-                                                          final Map<String, String> query,
-                                                          CMACallback<
-                                                              CMAArray<CMAContentType>> callback) {
+    public CMACallback<CMAArray<CMAContentType>> fetchAll(
+        final String spaceId,
+        final Map<String, String> query,
+        CMACallback<CMAArray<CMAContentType>> callback) {
       return defer(new DefFunc<CMAArray<CMAContentType>>() {
         @Override CMAArray<CMAContentType> method() {
-          DefaultQueryParameter.putIfNotSet(query, DefaultQueryParameter.FETCH);
           return ModuleContentTypes.this.fetchAll(spaceId, query);
         }
       }, callback);
@@ -380,11 +561,36 @@ public final class ModuleContentTypes extends AbsModule<ServiceContentTypes> {
      * @throws IllegalArgumentException if spaceId is null.
      * @throws IllegalArgumentException if contentTypeId is null.
      */
-    public CMACallback<CMAContentType> fetchOne(final String spaceId, final String contentTypeId,
-                                                CMACallback<CMAContentType> callback) {
+    public CMACallback<CMAContentType> fetchOne(
+        final String spaceId,
+        final String contentTypeId,
+        CMACallback<CMAContentType> callback) {
       return defer(new DefFunc<CMAContentType>() {
         @Override CMAContentType method() {
           return ModuleContentTypes.this.fetchOne(spaceId, contentTypeId);
+        }
+      }, callback);
+    }
+
+    /**
+     * Fetch a Content Type with the given {@code contentTypeId} from a Space.
+     *
+     * @param spaceId       Space ID
+     * @param environmentId Environment ID
+     * @param contentTypeId Content Type ID
+     * @param callback      Callback
+     * @return the given {@code CMACallback} instance
+     * @throws IllegalArgumentException if spaceId is null.
+     * @throws IllegalArgumentException if contentTypeId is null.
+     */
+    public CMACallback<CMAContentType> fetchOne(
+        final String spaceId,
+        final String environmentId,
+        final String contentTypeId,
+        CMACallback<CMAContentType> callback) {
+      return defer(new DefFunc<CMAContentType>() {
+        @Override CMAContentType method() {
+          return ModuleContentTypes.this.fetchOne(spaceId, environmentId, contentTypeId);
         }
       }, callback);
     }